Added riscv-tools git submodule.
Previously we have installed the toolchain from "some" version of the
riscv-gnu-toolchain repo. From now on we will recommend that the
toolchain be installed from the riscv-gnu-toolchain submodule in the
riscv-tools submodule.

This is because the toolchain needs to be compatible with the test suite
and the easiest way to ensure this is to depend on riscv-tools, wich has
compatible revisions of the toolchain and test repos as git submodules.