First revision of the riscv-tests simulator.
It was able to correctly detect that the rv32ui-p-simple test passed and
stop simulating at the correct moment. It does this by looking at the
stream of instructions going from fetch to decode and halting if it
discovers an instruction that is uniquely in TEST_PASS or
TEST_FAIL.

This technique is a hacky workaround for the fact that we haven't
impelemented the privileged instructions part of the ISA spec. yet.