Can now choose between RAID0 RAM or normal RAM.
Before there was only support for using the RAM that is in RAID0. Now
one can choose between normal RAM and a RAID0 RAM.

The motivation for this is that in simulation the RAM content's waveform
becomes messed up for RAID0, but is very readable for normal RAM.

A better solution would have been to keep the HW the same but solved the
undebuggable RAM problem some other way. E.g. by reading out every RAM
value to some dummy array when in simulation or extending ones waveform
viewer to handle the RAID0 format.

But it was easier to solve it like this right now ...