Added inital work on a Dockerfile.
The Docker image is based on a Arch Linux image with yaourt
installed. So far the necessary Arch Linux packages and an initial
attempt at installing the RISC-V toolchain has been made. But this is
still a WIP.

Remaining work to be done before this can be merged is installing
Xilinx and fixing the RISC-V toolchain.