Optimize docker image size
Deleted the riscv toolchain after building from source, reducing
images size from 8.5GB to 2.92GB.

Turborav utilizes a specific revision of the riscv toolchain
in order to avoid constant updates, as recompiling is a long
and tedious process. As such it is pointless to keep these
sources in the image due to the low frequency of rebuilds.