Big scary processor change.
Unfortunately, multiple things in the processor change with this commit.

Flushing pipelines is done in a more consistent way.
If the pipeline needs to be flushed (e.g. b/c of a branch
misprediction) a stage needs to signal that to the hdu,
and then the hdu will pass that signal on to the appropriate
stages.

Also some minor changes;
instead of a pc_sel signal (Select program counter) being sent
from exe, a branch_taken signal is sent. Semantics remain the
same.