Moved the register bank from decode into execute.
Also, made the reads to the register bank synchronous instead of
combinatorial. Moving from combinatorial to synchronous was done so that
FPGA block RAM could be used for the register bank memory, saving us 2k
LUTs.

Moving the register bank into execute was done because that was more
natural after having made the register file synchronous.