exe: Fix for the bug exposed by the hybrid test
Previously, when the execute stage received a stall and a flush signal
simultaneously the flush signal would have precedence. Now the stall
signal has precedence and the flush signal is ignored until the stage is
done stalling.

This precedence makes sense because the contract for stalling is that
you should ignore your inputs and retain your state until the stalling
is over.

The specific case that provoked the bug here was if we had a branch
taken simultaneously with a stall (the stall happened to come from a
multi-cycle memory transfer in this case).

It is execute itself that detects branch taken, and it would correctly
send a flush signal to itself and decode. But fetch would be stalled by
the memory transfer, and would therefore not receieve the branch taken
signal. And execute would flush itself, thereby losing the branch taken
signal. So the branch taken signal would get lost.

With this change execute does not flush itself, but instead holds the
branch taken signal until it is no longer stalled, at which point the
fetch signal receives the branch taken signal and we get the correct
behaviour.

Bah, this pipeline business is really messy. There has to be a way to
design pipelines without relying so much on testing, but have the code
be obviously correct.