Fixed clock frequency performance issue
During synthesis it was found that the writeback stage was
combinatorially using a signal from the memory stage. So effectively, we
didn't have a 5 stage pipeline, we had a 4 stage pipeline.

A major restructuring has been done to fix this perforfmance issue.
Changes include:

Memory now keeps track of the need to stall itself without relying on
the Hazard Detection Unit.

The fetch and memory stages now do synchronous reads of the ROM, instead
of asynchronous.

In addition there is a a lot more unit testing coverage of the Memory
stage now.