Optimization: Re-arrange HW to improve clk freq
Move mux between pc_next and alu_result from wrb to Mem.

This was done for clock frequency performance reasons. This mux was on
the critical path.

This optimization improves both the clock frequency and the area usage
because we the pipeline register size between mem and wrb has been
decreased.