csmith: Added support for Csmith based testing
I am trying to do some more extensive testing of the CPU. I had a look
at riscv-torture, but found that it was a hassle to set up for
RV32UI[0]. So instead I am copying Clifford Wolf's technique of
generating C programs and verifying that the custom CPU executes these
programs in the same way as a golden model does.

Not much is supported yet;

Csmith has been added as a dependency in the Dockerfile

The Csmith test harness has been modified slightly to be compatible with TurboRav.

The build system can generate random C programs with Csmith, build them
for the host machine, and also build and simulate them on TurboRav.

Next up is;

Support for timeout of Csmith programs.

Support for checking that TurboRav computes the same as the host
machine (or maybe spike if the C programs are architecture-dependent).

I have simulated 16 programs so far, the fifth program found a bug in
the interaction between the linker script and the startup assembly code.
To simulate more programs, I am going to need automatic checking of the
C programs results.