fetch: Resolve structural hazard over ROM access
There is a structural hazard over the ROM resource. Both Fetch and
Memory want access to it, but it can only serve one request per clock
cycle. We give Memory the priority to ROM because the instruction in
Fetch is more likely to be flushed than the instruction in Memory.

This means that fetch must stall when the ROM is busy from a Memory
request. Fetch can detect that the ROM is busy with the
fch.io.rr_io.response.valid signal.