linker_script: Align sections on 4 byte boundaries
This makes it easier to copy sections from FW e.g. when initializing the
.data section we can use only 4-byte lw and sw instructions.

Previously, if we had a data section that wasn't aligned on a 4 byte
boundary, the startup code that copies the section would break, because
it assumed the sections were aligned.

This linker script / startup code bug was found by a randomly generated
csmith program with seed 5!