rtl: Fixed bug provoked by csmith seed 35
Multiple changes were needed to fix this bug. The underlying issue was
that when the exe stage was stalled it did not kill all of it's outgoing
signals. This caused exe to corrupt other stages even though it should
have been NOPing from a stall.

Specifically, exe was firing on the branch_taken signal and this was
causing instructions to be flushed due to branch misprediction.

Now exe kills it's branch_taken signal when it stalls. To make this
change without introducing a combinatorial loop I also had to ensure
that exe did not stall itself. Exe is now aware of stalls due to to
external factors and due to internal factors and acts slightly
differently in these two situations.